requirements.txt
pdfplumber>=0.7.7
pandas>=1.3
python-dateutil>=2.8.2
regex
tqdm
# optional for scanned PDFs / OCR (only if needed)
pytesseract>=0.3.10
Pillow>=9.0
